ESDEV-3641 Remove unrelated functionality from OxidEsales\Eshop\Core\Database. Add exception handling to OxidEsales\Eshop\Core\ShopControl
The Database class should only be responsible  for creating and providing a database connection.
Therefor the exception handling which took place in the class was mover to the ShopControl class.
A new Exception class OxidEsales\Eshop\Exception\DatabaseNotConfiguredException was created in order to be able to handle the fact that the shop might not have been configured yet.
Unused code was removed from the Database class.
Test were adapted.