ESDEV-3641 Remove  escapeString() from Database class.
This method was used in very different contexts. According to the context a proper solution for replacing the method was applied.
In some places Database::getDb->quoteArray() was the choice, in some places it was Database::getDb->quoteIdentifier(), in some places it was Database::getDb->quote(). In some place a prepared statement could be used.